Added dummy implementation of OS_PlatformFeatures 0, 33 and 64.
OS_PlatformFeatures 0 to read the code features is now returning
some values - all the values actually in 32bit, and a restricted
set for 64bit.

OS_PlatformFeatures 33 returns some dummy values for the cache
details.

OS_PlatformFeatures 64 returns a static table for 32bit and 64bit
in an ugly static table of words.